# docs/swagger.yaml
basePath: /
definitions:
  structure.Announcement:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      message:
        type: string
      priority:
        type: string
      started_at:
        type: string
      type:
        type: string
    type: object
  structure.AuditTrail:
    properties:
      pushed_at:
        type: string
      pushed_by:
        type: string
      via:
        type: string
    type: object
  structure.Locale:
    enum:
    - en
    - zh
    type: string
    x-enum-varnames:
    - LocaleEN
    - LocaleZH
  structure.NotificationContent:
    properties:
      audit_trail:
        $ref: '#/definitions/structure.AuditTrail'
      channel:
        type: string
      image_url:
        type: string
      locale:
        $ref: '#/definitions/structure.Locale'
      message:
        type: string
      segments:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
info:
  contact: {}
  description: |-
    提供推送通知相关的 API 接口，支持发送文本和图片通知，集成 OneSignal 推送服务，
    提供 WebSocket 实时通知功能，支持向所有用户广播通知，
    同时提供通知记录存储和管理功能，为应用提供完整的推送通知解决方案
  title: 推送通知服务
  version: "1.0"
paths:
  /announcement/all:
    get:
      consumes:
      - application/json
      description: 获取所有公告信息，按创建时间降序排序
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有公告
      tags:
      - 公告管理
  /announcement/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的公告并保存到数据库
      parameters:
      - description: 公告信息
        in: body
        name: announcement
        required: true
        schema:
          $ref: '#/definitions/structure.Announcement'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建新公告
      tags:
      - 公告管理
  /announcement/delete:
    delete:
      consumes:
      - application/json
      description: 根据 ID 删除公告
      parameters:
      - description: 公告 ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除公告
      tags:
      - 公告管理
  /announcement/latest:
    get:
      consumes:
      - application/json
      description: 获取最新的公告信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 没有找到公告
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取最新公告
      tags:
      - 公告管理
  /announcement/update:
    put:
      consumes:
      - application/json
      description: 根据 ID 更新公告信息
      parameters:
      - description: 公告 ID
        in: query
        name: id
        required: true
        type: string
      - description: 公告信息
        in: body
        name: announcement
        required: true
        schema:
          $ref: '#/definitions/structure.Announcement'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新公告
      tags:
      - 公告管理
  /push/text:
    post:
      consumes:
      - application/json
      description: |-
        发送通用文本推送通知到所有用户。此端点用于发送面向全体用户的通知，如系统公告、重要更新等。
        警告：请勿使用此端点发送用户特定的通知，因为它会广播给所有用户。对于用户特定的通知，请使用专门的用户定向通知端点。
      parameters:
      - description: 通知内容
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/structure.NotificationContent'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 发送文本推送通知
      tags:
      - 通知
  /push/text-image:
    post:
      consumes:
      - application/json
      description: |-
        发送包含图片的推送通知到所有用户。此端点用于发送面向全体用户的通知，如系统公告、重要更新等，支持添加图片。
        警告：请勿使用此端点发送用户特定的通知，因为它会广播给所有用户。对于用户特定的通知，请使用专门的用户定向通知端点。
      parameters:
      - description: 通知内容
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/structure.NotificationContent'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 发送文本和图片推送通知
      tags:
      - 通知
  /segment/all:
    get:
      consumes:
      - application/json
      description: 获取 OneSignal 中的所有分段信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有分段
      tags:
      - 分段管理
swagger: "2.0"